backend/services/transcription/requirements.txt
pynamodb~=5.5.1
boto3~=1.34.11
botocore~=1.34.11
requests~=2.31.0
beautifulsoup4~=4.12.2
fastapi~=0.108.0
pydantic~=2.5.3
uvicorn~=0.25.0
python-dotenv~=1.0.0